blog_os/requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.5.0
django-environ==0.4.5
django-summernote==0.8.11.6
djangorestframework==3.12.1
Pillow==8.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.4.1
